Implement full CRUD for products

+++ backend/app/db/models/product_model.py
@@ -1,5 +1,5 @@
 
-from sqlalchemy import Table, Column, Integer, String, DECIMAL, ForeignKey
+from sqlalchemy import Table, Column, Integer, String, ForeignKey, Float
 from backend.app.config.db import meta, engine
 from backend.app.db.models.category_model import categories
 
@@ -8,7 +8,7 @@ products = Table("products", meta,
                        Column("name", String(255), nullable=False),
                        Column("description", String(255), nullable=False),
                        Column("stock", Integer, nullable=False),
-                       Column("price", DECIMAL, nullable=False),
+                       Column("price", Float, nullable=False),
                        Column("category_id", Integer, ForeignKey('categories.id'))
                  )
